[Naxx 2] Extract Era Naxx config
So that we can make changes for SoD in a follow up PR
